Fix gcc warning: Tests/CachedHashTableTest.cpp:654:5: required from here ThirdParty/googletest/gtest/gtest.h:10751:11: warning: comparison of integer expressions of different signedness: ‘const long unsigned int’ and ‘const long int’ [-Wsign-compare] 10751 | if (lhs == rhs) { | ~~~~^~~~~~

+++ Tests/CachedHashTableTest.cpp
@@ -652,7 +652,7 @@ TEST(Truncate, OverlapsJoinCacheInvalidationTest) {
     auto row = result->getNextRow(false, false);
     EXPECT_EQ(size_t(1), row.size());
     auto count = boost::get<int64_t>(boost::get<ScalarTargetValue>(row[0]));
-    EXPECT_EQ(size_t(2), count);
+    EXPECT_EQ(2, count);
   }
   EXPECT_EQ(QR::get()->getNumberOfCachedOverlapsHashTables(),
             size_t(2));  // bucket threshold and hash table
@@ -676,7 +676,7 @@ TEST(Truncate, OverlapsJoinCacheInvalidationTest) {
     auto row = result->getNextRow(false, false);
     EXPECT_EQ(size_t(1), row.size());
     auto count = boost::get<int64_t>(boost::get<ScalarTargetValue>(row[0]));
-    EXPECT_EQ(size_t(1), count);
+    EXPECT_EQ(1, count);
   }
   EXPECT_EQ(QR::get()->getNumberOfCachedOverlapsHashTables(),
             size_t(1));  // bucket threshold and hash table
